# perf(language-service): improve Language service performance