--- a/packages/compiler/src/aot/static_symbol_resolver.ts
+++ b/packages/compiler/src/aot/static_symbol_resolver.ts
@@ -273,9 +273,7 @@ export class StaticSymbolResolver {
     // Note: Some users use libraries that were not compiled with ngc, i.e. they don't
     // have summaries, only .d.ts files, but `summaryResolver.isLibraryFile` returns true.
     this._createSymbolsOf(filePath);
-    const metadataSymbols = this.symbolFromFile.get(filePath) || [];
-    const uniqueMetadataSymbols = new Set<StaticSymbol>(metadataSymbols);
-    return Array.from(uniqueMetadataSymbols);
+    return this.symbolFromFile.get(filePath) || [];
   }
 
   private _createSymbolsOf(filePath: string) {
@@ -359,9 +357,12 @@ export class StaticSymbolResolver {
             this.createResolvedSymbol(symbol, filePath, topLevelSymbolNames, symbolMeta));
       });
     }
-    resolvedSymbols.forEach(
-        (resolvedSymbol) => this.resolvedSymbols.set(resolvedSymbol.symbol, resolvedSymbol));
-    this.symbolFromFile.set(filePath, resolvedSymbols.map(resolvedSymbol => resolvedSymbol.symbol));
+    const uniqueSymbols = new Set<StaticSymbol>();
+    for (const resolvedSymbol of resolvedSymbols) {
+      this.resolvedSymbols.set(resolvedSymbol.symbol, resolvedSymbol);
+      uniqueSymbols.add(resolvedSymbol.symbol);
+    }
+    this.symbolFromFile.set(filePath, Array.from(uniqueSymbols));
   }
 
   private createResolvedSymbol(