## PR Checklist
Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: https://github.com/angular/angular/blob/master/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)


## PR Type
What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] angular.io application / infrastructure changes
- [ ] Other... Please describe:


## What is the current behavior?
A lot of unnecessary calculations is doing in `StaticSymbolResolver.getSymbolsOf` method. Metadata symbols for file can be retrieved from `symbolFromFile` map instead of traversing all symbols.
![profiler](https://user-images.githubusercontent.com/17968027/62836054-f041ac80-bc67-11e9-824d-49d6e034b033.PNG)
![profiler2](https://user-images.githubusercontent.com/17968027/62836057-f2a40680-bc67-11e9-9568-c549d29d4260.PNG)
The average response times are  for completions - 342ms and for diagnostics - 117ms. The project has 113 Html files, 1114 ts files.

Issue Number: N/A


## What is the new behavior?
The average response times for completions is 156ms and for diagnostics - 79ms.

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No


<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->


## Other information