@kyliau 2019-09-07
@alxhub, the change makes sense to me, could you please take a look? Thank you!